# fastfood-operations/internal/orders/adapters/http/requests/update_order_request.go
package requests

import (
	"github.com/SOAT-46/fastfood-operations/internal/orders/domain/entities"
)

type UpdateOrderRequest struct {
	Status string `json:"status" binding:"required"`
}

func (itself UpdateOrderRequest) ToDomain(id int) entities.Order {
	status := entities.ToOrderStatus(itself.Status)
	return entities.Order{
		ID:     id,
		Status: status,
	}
}